pyproject.toml
[tool.poetry]
name = "techcommerce"
version = "0.1.0"
description = "MVP e‑commerce platform built with Django"
authors = ["Fernando Marques <fernando@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
psycopg2-binary = "^2.9"
Faker = "^25.0"
gunicorn = "^22.0"
dj-database-url = "^2.2"
python-decouple = "^3.8"
django-filter = "^25.1"

[tool.poetry.dev-dependencies]
pytest = "^8.0"
pytest-django = "^4.7"
pytest-cov = "^5.0"
black = "^24.4"
isort = "^5.13"
mypy = "^1.9"

[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=. --cov-config=.coveragerc --cov-report=term-missing"

